Se han agregado animaciones para el movimiento horizontal y el vertical del personaje. Se ha corregido un error entre el salto doble y el coyotetime donde si saltabas haciendo uso del coyotetime se consumía el primer salto extra.
