+++ TFG Plataformer/Assets/Animation/Player/Player_fall.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c59d7be7c1b79054a886a154fa03f890
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
