+++ TFG Plataformer/Assets/Scripts/Player/PlayerMovementTemplate.cs
@@ -1,0 +1,77 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerMovementTemplate : MonoBehaviour {
+
+    public CharacterController2D controller;
+    public Animator animator;
+
+    public static float runSpeed = 40f;
+
+    float horizontalMove = 0f;
+    bool jump = false;
+    bool crouch = false;
+
+    public GameObject[] players;
+
+
+    void Start()
+    {
+        DontDestroyOnLoad(gameObject);
+    }
+    // Update is called once per frame
+    void Update()
+    {
+        horizontalMove = Input.GetAxisRaw("Horizontal") * runSpeed;
+        
+        animator.SetFloat("Speed", Mathf.Abs(horizontalMove));
+
+        if(Input.GetButtonDown("Jump"))
+        {
+            jump = true;
+            animator.SetBool("IsJumping", true);
+        }
+
+        if (Input.GetButtonDown("Crouch"))
+        {
+            crouch = true;
+        } else if (Input.GetButtonUp("Crouch")) 
+        { 
+            crouch = false;
+        }
+
+    }
+
+    public void OnLanding ()
+    {
+        animator.SetBool("IsJumping", false);
+    }
+
+    public void OnCrouching (bool isCrouching)
+    {
+        animator.SetBool("IsCrouching", isCrouching);
+    }
+    void FixedUpdate ()
+    {
+        controller.Move(horizontalMove * Time.fixedDeltaTime, crouch, jump);
+        jump = false;
+
+    }
+
+    private void OnLevelWasLoaded(int level)
+    {
+        FindStartPos();
+        players = GameObject.FindGameObjectsWithTag("Player");
+
+        if(players.Length > 1)
+        {
+            Destroy(players[1]);
+        }
+    }
+
+    void FindStartPos()
+    {
+        transform.position = GameObject.FindWithTag("StartPos").transform.position;
+    }
+}
